Creates a new trait and vfs+REST test case.
