+++ Integration/RESTVfsTestCase.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace WPMedia\PHPUnit\Integration;
+
+/**
+ * RESTful Virtual Filesystem Test Case.
+ */
+abstract class RESTVfsTestCase extends VirtualFilesystemTestCase {
+	use RESTTrait;
+
+	public function setUp() {
+		parent::setUp();
+
+		$this->setUpServer();
+	}
+}
