+++ Integration/RESTfulTestCase.php
@@ -9,12 +9,7 @@ use WP_UnitTestCase;
 
 abstract class RESTfulTestCase extends TestCase {
 	use ApiTrait;
-
-	/**
-	 * Instance of the WordPress REST Server.
-	 * @var WP_REST_Server
-	 */
-	protected $server;
+	use RESTTrait;
 
 	/**
 	 * Name of the API credentials config file, if applicable. Set in the test or new TestCase.
@@ -25,34 +20,9 @@ abstract class RESTfulTestCase extends TestCase {
 	 */
 	protected static $api_credentials_config_file;
 
-	/**
-	 * Setup the WP REST API Server.
-	 */
 	public function setUp() {
 		parent::setUp();
-		/**
-		 * @var WP_REST_Server $wp_rest_server
-		 */
-		global $wp_rest_server;
-		$this->server = $wp_rest_server = new WP_REST_Server;
-		do_action( 'rest_api_init' );
-	}
 
-	/**
-	 * Does the REST request.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param array $body_params Body parameters.
-	 * @param  string     $route Requested route.
-	 *
-	 * @return WP_REST_Response REST response.
-	 */
-	protected function doRestRequest( array $body_params, $route ) {
-		$request = new WP_Rest_Request( 'PUT', $route );
-		$request->set_header( 'Content-Type', 'application/x-www-form-urlencoded' );
-		$request->set_body_params( $body_params );
-
-		return rest_do_request( $request )->get_data();
+		$this->setUpServer();
 	}
 }
